Added moderation
Added some navigation where there now is a choice between user and moderator. This is not a safe navigation because you do not need any credentials to act as a moderator. Also collected all moderation data that is combined with the message data read from a csv file and sent off to the database.