Improved GUI in looks and fuctionality.
Text entry fonts now match the rest of the style. The NS logo at the top is now a button that brings you to the start. Sending a message in the GUI now actually works in the back-end, same goes for approving a message. Weather API is now linked with the GUI and the utilities are accurately reflected per station.